Move functions to 'module.helpers' so they can be reused by other modules
